LU-15113 tests: Add margin to 398g
Every once in a great while, some other operation I can't
identify triggers a single write RPC to a different file
in test 398g on Gatekeeper testing.

This has nothing to do with the test itself, but does
cause it to fail occasionally.  An easy solution that
isn't too bad for the test is to add a margin of +1 RPCs
to account for this.

Only modifies sanity, so trivial is OK.

test-parameters: trivial

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I3936077cb60259653628ed26b01470ff529b0272
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/45260
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Vitaliy Kuznetsov <vkuznetsov@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>