LU-15277 utils: lfs quota/setquota improvements
Add long options to "lfs quota" for ease of use.  Improve usage
message for "lfs quota" and "lfs setquota" to match current code.

Deprecate the "lfs quota -i MDT_IDX|-I OST_IDX" options to print one
target, since these arguments are backward from other lfs subcommands.
Add "-o" for the OST_IDX and "-m" for the MDT_IDX and long options
"--ost" and "--mdt" to match other lfs subcommands.  We may eventually
be able to liberate "-i" to use the OST_IDX, but not for a while yet.

Fix "lfs setquota" handling of long --times option.  It was being
checked in lfs_setquota_times(), but not in has_times_option().

Sort arguments to be handled (as much as possible) in alphabetical
order for ease of use in the future.

Update lfs-quota.1 and lfs-setquota.1 man pages to describe all
options and add proper argument formatting.

Test-Parameters: trivial testlist=sanity-quota
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I049f22a526469ea1ed1da04beffda6bb683ebbe6
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/46615
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sergey Cheremencev <scherementsev@ddn.com>
Reviewed-by: Hongchao Zhang <hongchao@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>