LU-15619 osc: Remove oap lock
The OAP lock is taken around setting the oap flags, but not
any of the other fields in oap.  As far as I can tell, this
is just some cargo cult belief about locking - there's no
reason for it.

Remove it entirely.  (From the code, a queued spin lock
appears to be 12 bytes on x86_64.)

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: Ib61190d52c08d88c95a0c19b8ef7d114e26cfae2
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/46719
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Zhenyu Xu <bobijam@hotmail.com>