LU-10499 pcc: Limit attach queue depth
The existing async attach code does not attempt to limit
the number of async attaches that can be requested at once.
This is a problem because we could theoretically create too
many kthreads and overwhelm the system.

When the attach queue depth is exceeded, we stop allowing
new items to be queued by switching over to sync attach.

Ideally we would rebuild the attach code to generate a
queue of attach requests and have the attach thread code
pull items from the queue until it's exhausted, but that's
a much more substantial change and is left for later.

It's really hard to remember what 'inode' and 'file' mean
when there's more than one in a function, so I've redone
some of the names here.

EX-6399 pcc: add tunable parameter for PCC attach thread

Add a tunable parameter llite.*.pcc_attach_thread_max to
tune the max number of kernel threads doing asynchronous
attach (1024 by default).

Was-Change-Id: Ic59c15af935dd8dff586fa6be3939d4322c136d5

Use variable names that are consistent with parameter name.
Add llite.pcc_attach_thread_max.4 man page for parameter.

Fixes: 5be466e87304 ("LU-10499 pcc: async attach in the background for PCC-RO file")
EX-bug-id: EX-5014 EX-6399
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Signed-off-by: Qian Yingjin <qian@ddn.com>
Signed-off-by: Andreas Dilger <adilger@thelustrecollective.com>
Change-Id: Ib00dfb67f5245a28b722278d031ee8cdf5e190d6
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54476
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>